split main file, one per command reuse code when possible through sub commands (install, create, update)
