+++ src/micromamba/options.cpp
@@ -1,0 +1,21 @@
+// Copyright (c) 2019, QuantStack and Mamba Contributors
+//
+// Distributed under the terms of the BSD 3-Clause License.
+//
+// The full license is in the file LICENSE, distributed with this software.
+
+#include "options.hpp"
+
+#include <string>
+#include <vector>
+
+
+ShellOptions shell_options{};
+CleanOptions clean_options{};
+CreateOptions create_options{};
+UpdateOptions update_options{};
+NetworkOptions network_options{};
+ConfigOptions config_options{};
+ListOptions list_options{};
+GeneralOptions general_options{};
+ConstructorOptions constructor_options{};
